Extract class: AzureCli
